:recycle: Refactored code
May cause regressions :grimacing:
